pyproject.toml
[project]
name = "asynchronous-django-ecommerce-api-with-psycopg3-and-uvicorn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "adrf>=0.1.9",
    "celery[redis]>=5.5.2",
    "django>=5.2.1",
    "django-debug-toolbar>=5.2.0",
    "django-filter>=25.1",
    "django-redis>=5.4.0",
    "django-taggit>=6.1.0",
    "djangorestframework-camel-case>=1.4.2",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "pillow>=11.2.1",
    "psycopg[async,binary]>=3.2.9",
    "pydantic-settings>=2.9.1",
    "python-slugify>=8.0.4",
    "uvicorn[standard]>=0.34.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "django-debug-toolbar>=5.2.0",
]
